+++ examples/data-sources/nxos_loopback_interface_vrf/data-source.tf
@@ -1,0 +1,3 @@
+data "nxos_loopback_interface_vrf" "example" {
+  interface_id = "lo123"
+}