+++ examples/resources/nxos_loopback_interface_vrf/import.sh
@@ -1,0 +1,1 @@
+terraform import nxos_loopback_interface_vrf.example "sys/intf/lb-[lo123]/rtvrfMbr"