+++ examples/resources/nxos_loopback_interface_vrf/resource.tf
@@ -1,0 +1,4 @@
+resource "nxos_loopback_interface_vrf" "example" {
+  interface_id = "lo123"
+  vrf_dn       = "sys/inst-default"
+}