+++ examples/resources/nxos_svi_interface_vrf/resource.tf
@@ -1,4 +1,4 @@
 resource "nxos_svi_interface_vrf" "example" {
   interface_id = "vlan293"
-  vrf_dn       = "sys/inst-default"
+  vrf_dn       = "sys/inst-VRF123"
 }